fix: review Wave 1 — correctness (deep-link views, optical leak, error class, grammar, f32 axis)
From the 2026-06-15 adversarial review (.planning/research/CODE-REVIEW-2026-06-15.md):

- A1: add `overview` + `multi` to the URL grammar's VALID_VIEWS — those imaging views couldn't
  be deep-linked and `serialize(view=multi)` didn't round-trip. +round-trip tests.
- A4: `ion=mz,` (empty tol) now defaults to DEFAULT_TOL_DA instead of `Number("")===0` (a
  zero-width window → blank image); channel `color` may contain commas (`rgb(1,2,3)`); `xic`
  empty delta rejected. +tests.
- A5: the mean/ROI reference axis is now f32 CONSISTENTLY (cold f64 source downcast on capture)
  so a mean spectrum is identical warm-vs-cold; adaptSpectrum widens to f64 for the wire.
  +assertions (cold golden + warm cache are both f32-representable).
- A3: wire CorruptFileError + tag network errors at the open boundary — the "parse"/"network"
  classes were unreachable (only requireActive set engineClass), so corrupt files and fetch
  failures both showed "internal".
- A2: the optical-image gen guard was DEAD (never incremented) and the decode caches leaked
  across imaging→imaging file switches — added a reset effect keyed on opticalImages.
- A6: the auto-expand-accordion setter was scheduled from render via Promise.then (setState in
  render) — moved to a useEffect.

142 core (+1) / 53 contracts (+4) / 19 ui-kit + 18 e2e green; typecheck + build clean.

+++ packages/core/src/engine/imaging.ts
@@ -562,14 +562,18 @@ const MAX_SAMPLES = 300;
 const MAX_ROI = 100;
 
 type MeanAccumulator = {
-  refMz: Float64Array | null;
+  // f32 reference axis CONSISTENTLY (the cold f64 source is downcast on capture), so a
+  // mean/ROI spectrum is identical whether built from the f32 warm cache or a cold f64 read —
+  // matching the f32 ion pipeline. (adaptSpectrum widens it to f64 for the wire; the VALUES
+  // are f32-precision either way.) Intensity sum stays f64 for accumulation accuracy.
+  refMz: Float32Array | null;
   intensitySum: Float64Array | null; // per-bin running sum (indexed by ref bin)
   countPerBin: Int32Array | null; // per-bin contributing-spectrum count
   contributed: number;
 };
 
 /** Nearest reference-bin index for `mzVal` via binary search (IV inner loop). */
-function nearestBin(refMz: Float64Array, mzVal: number): number {
+function nearestBin(refMz: Float32Array, mzVal: number): number {
   let lo = 0;
   let hi = refMz.length - 1;
   while (lo < hi) {
@@ -592,9 +596,9 @@ function accumulate(
   if (n === 0) return;
 
   if (acc.refMz === null) {
-    // First spectrum defines the reference axis (sorted ascending — IV sorts here;
-    // engine/spectrum.ts already returns sorted, finite pairs, so this is stable).
-    const ref = new Float64Array(n);
+    // First spectrum defines the reference axis (sorted ascending). f32 axis: an f64 source
+    // is downcast here so the axis is f32-precision regardless of cache state (see type doc).
+    const ref = new Float32Array(n);
     const sum = new Float64Array(n);
     const cnt = new Int32Array(n);
     for (let i = 0; i < n; i++) {
